include ps -au in if statement

+++ src/test_clustering.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+i#!/bin/sh
 
 ################################################################################
 #! Initialize shell environment
@@ -119,6 +119,8 @@ cat > "${query_label}"
 ################################################################################
 
 echo "Clustering ${barcode} ${alleletype}..." >&2
-[ "$(cat ${query_label} | wc -l)" -gt 50 ] &&
+if [ "$(cat ${query_label} | wc -l)" -gt 50 ]; then
     Rscript DAJIN/src/test_clustering.R "${query_score}" "${query_label}" "${control_RDS}" "${threads}" 2>&1 | grep -v "joblib" #2>/dev/null || echo "TERMINATED" #! RENAME =================
-ps -au | grep -e "clustering.R" -e "joblib" | awk '{print $2}'| xargs kill 2>/dev/null || true
+    ps -au | grep -e "clustering.R" -e "joblib" | awk '{print $2}'| xargs kill 2>/dev/null || true
+fi
+
